requirements.txt
# SẮC XUÂN VĂN HỌC - Requirements
# Python dependencies

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1
Flask-Compress==1.23

# WSGI Server (for production)
gunicorn==21.2.0

# Optional: AI/ML Libraries (for future integration)
# OpenAI SDK (used by /api/generate-poem and /api/generate-image)
openai>=1.0.0

# Load .env locally
python-dotenv>=1.0.0

# HTTP client (optional fallback)
requests>=2.31.0
# anthropic==0.7.8
# langchain==0.1.0
# transformers==4.36.0
# torch==2.1.0
# pillow==10.1.0
